+++ api/apps/test/Test.ts
@@ -1,4 +1,4 @@
-import { AppSpec } from '../../services/app/App'
+import { AppSpec } from '../../app-base/App'
 
 async function run({ action, env, params: { path } }) {
   return { type: 'embed', path: `/tests/${path}` }
