+++ api/channel/ConsoleChannel.js
@@ -1,12 +1,12 @@
 import shortid from 'shortid'
 import ClientFileStore from '../storage/ClientFileStore'
-import ConsoleError from '../../ConsoleError'
-import App from '../app/App'
-import parseArgs from '../app/parseArgs'
-import parseUrl from '../app/parseUrl'
-import Asana from '../../apps/asana/Asana'
-import GitHub from '../../apps/github/GitHub'
-import Test from '../../apps/test/Test'
+import ConsoleError from '../ConsoleError'
+import App from '../app-base/App'
+import parseArgs from '../app-base/parseArgs'
+import parseUrl from '../app-base/parseUrl'
+import Asana from '../apps/asana/Asana'
+import GitHub from '../apps/github/GitHub'
+import Test from '../apps/test/Test'
 import env from './env'
 
 const apps = {
