+++ api/storage/BrowserFileStore.js
