+++ pages/api/channels/[name]/files/[...path].js
@@ -1,7 +1,7 @@
-import ConsoleError from 'console/ConsoleError'
-import ConsoleWorkspace from 'console/services/workspace/ConsoleWorkspace'
-import ConsoleChannel from 'console/services/channel/ConsoleChannel'
-import LocalFileStore from 'console/services/storage/LocalFileStore'
+import ConsoleError from 'api/ConsoleError'
+import ConsoleWorkspace from 'api/workspace/ConsoleWorkspace'
+import ConsoleChannel from 'api/channel/ConsoleChannel'
+import LocalFileStore from 'api/storage/LocalFileStore'
 import { isObject } from 'vtv/model/analyze'
 
 ConsoleWorkspace.LocalFileStore = LocalFileStore
